[CHORE] CustomTextButton로 버튼 통합 및 CommonComposable 함수 분리
- SimpleNarrowTextButton와 SimpleWideTextButton를 CustomTextButton으로 변경 및 적용
- CommonComposable에 있는 SimpleTextAndIcon과 CustomTextField를 분리
- CommonComposable을 SimpleTextAndIcon으로 이름 변경

Related to: TICO-31
