+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/common/view/CustomTextField.kt
@@ -1,0 +1,87 @@
+package com.tico.pomorodo.ui.common.view
+
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.TextFieldColors
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.material3.TextFieldDefaults.indicatorLine
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CustomTextField(
+    modifier: Modifier = Modifier,
+    value: String,
+    onValueChange: (String) -> Unit,
+    callback: (() -> Unit)? = null,
+    enabled: Boolean = true,
+    isError: Boolean = false,
+    singleLine: Boolean = true,
+    colors: TextFieldColors,
+    placeholder: @Composable (() -> Unit)? = null,
+    trailingIcon: @Composable (() -> Unit)? = null,
+    shape: Shape = TextFieldDefaults.shape,
+    contentPadding: PaddingValues = TextFieldDefaults.contentPaddingWithoutLabel(
+        start = 2.dp, top = 0.dp, end = 2.dp, bottom = 3.dp
+    )
+) {
+    val interactionSource = remember { MutableInteractionSource() }
+    val keyboardController = LocalSoftwareKeyboardController.current
+
+    BasicTextField(
+        value = value,
+        onValueChange = onValueChange,
+        modifier = modifier
+            .indicatorLine(
+                enabled = enabled,
+                isError = false,
+                interactionSource = interactionSource,
+                colors = colors
+            ),
+        visualTransformation = VisualTransformation.None,
+        interactionSource = interactionSource,
+        enabled = enabled,
+        keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done),
+        keyboardActions = KeyboardActions(
+            onDone = {
+                callback?.invoke()
+                keyboardController?.hide()
+            },
+        ),
+        singleLine = singleLine
+    ) { innerTextField ->
+        TextFieldDefaults.DecorationBox(
+            value = value,
+            visualTransformation = VisualTransformation.None,
+            innerTextField = innerTextField,
+            enabled = enabled,
+            placeholder = placeholder,
+            singleLine = singleLine,
+            colors = colors,
+            interactionSource = interactionSource,
+            contentPadding = contentPadding,
+            isError = isError,
+            container = {
+                TextFieldDefaults.Container(
+                    enabled = enabled,
+                    isError = isError,
+                    interactionSource = interactionSource,
+                    colors = colors,
+                    shape = shape,
+                )
+            },
+            trailingIcon = trailingIcon
+        )
+    }
+}
